fix components structure
