+++ client/src/containers/analysis-visualization/analysis-table/table/component.tsx
