+++ client/src/containers/analysis-visualization/analysis-table/table/index.ts
@@ -1,0 +1,1 @@
+export { default } from './component';
